split half page letter classes into two files
