+++ planner/python/classes/page_entries/free_write_entry.py
@@ -31,7 +31,7 @@ import svgwrite.container
 from classes.constants.addl_arg_keys import AddlArgKeys as Key
 from classes.elements.row_group import DualLineRowGroup
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 
 
 #_______________________________________________________________________
