+++ planner/python/classes/page_entries/free_write_prompt_entry.py
@@ -35,7 +35,7 @@ from classes.elements.row_group import DualLineRowGroup
 from classes.style.std_styles import StdTextBoxStyles
 from classes.style.style import PlannerFontStyle as Font
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 
 
 #_______________________________________________________________________
