+++ planner/python/classes/page_entries/test_entry.py
@@ -42,7 +42,7 @@ from classes.elements.table import DualLineTable
 from classes.elements.table import SingleLineTable
 from classes.elements.table import ColumnTable
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 
 
 #_______________________________________________________________________
