+++ planner/python/classes/page_entries/week_checklist_entry.py
@@ -33,7 +33,7 @@ from classes.elements.table import ColumnTable
 from classes.elements.table import DualLineTable
 from classes.style.std_styles import StdTextBoxStyles
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 
 #_______________________________________________________________________
 class WeekCheckList(OnePageHalfLetterLayout):
