+++ planner/python/classes/page_entries/week_habit_entry.py
@@ -37,7 +37,7 @@ from classes.elements.table import DualLineTable
 from classes.elements.table import SingleLineTable
 from classes.style.std_styles import StdTextBoxStyles
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 
 #_______________________________________________________________________
 class HabitTracker(OnePageHalfLetterLayout):
