+++ planner/python/classes/reference_pages/ace_reference.py
@@ -32,7 +32,7 @@ from classes.constants.addl_arg_keys import AddlArgKeys as Key
 from classes.constants.strings import PlannerStrings as Strings
 from classes.elements.row_group import DualLineRowGroup
 
-from classes.page_layouts.half_letter_layout import OnePageHalfLetterLayout
+from classes.page_layouts.half_letter_one_page import OnePageHalfLetterLayout
 from classes.elements.row_group import TextRowGroup
 from classes.style.std_styles import StdTextBoxStyles
